refactor: 共有ライブラリをmotionpngtuber/パッケージに整理
- 12個の共有ライブラリモジュールをルートからmotionpngtuber/パッケージに移動
- 全import文をmotionpngtuber.xxx形式に更新（スクリプト・テスト・mouth_track_gui）
- pyproject.tomlにpytest設定（testpaths, pythonpath）とdev依存（pytest）を追加
- mouth_sprite_extractor.pyはルートにCLIエントリポイントとして残存
- 全254テスト通過確認済み
