+++ calibrate_mouth_track.py
@@ -36,8 +36,8 @@ from dataclasses import dataclass
 import cv2
 import numpy as np
 
-from image_io import read_image_bgra
-from mouth_color_adjust import (
+from motionpngtuber.image_io import read_image_bgra
+from motionpngtuber.mouth_color_adjust import (
     MouthColorAdjust,
     apply_inspect_boost_3ch,
     apply_mouth_color_adjust_4ch,
