+++ motionpngtuber/auto_crop_estimator.py
