+++ motionpngtuber/realtime_emotion_audio.py
